Add git as a requirement
Many of the projects which rely on diskimage-builder pull down
sources from a Git repository for the majority of the time.  By
default, Git is not a requirement to DIB therefore the user has
to install the package manually.

This package includes it as a requirement which will lead to an
easier user experience, avoiding manually installing extra
packages to get an image build going.

Change-Id: Ib3b02eea77f4b4225f69f51de1baf8c47d8ac3b0