Don't depend on dib-utils
The original idea of dib-utils was to be a generic home for
interesting and re-usable tools that started in dib.  This never
really happened, and the only tool in there was dib-run-parts.
Eventually, dib wanted to make dib-specific changes to this and it
made no sense to do extra work making it all generic for a tool that
nobody is using.

Thus dib-run-parts was moved into dib's private library directory.
During v2 development there was an oversight that dib was putting
dib-run-parts into the global namespace, but that was corrected with
[1].

Thus there is no need to depend on dib-utils, and there is no need to
remove dib-run-parts as dib isn't installing it any more.

[1] https://review.openstack.org/#/c/446285/

Change-Id: I477db6b592a5e27f4019d51ac5386db61661ff17